Refactor alarmFragment name into alarm list fragment.
