+++ app/src/main/res/layout/fragment_alarm_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".main.alarm.AlarmFragment">
+    tools:context=".main.alarmList.AlarmListFragment">
 
     <TextView
         android:id="@+id/text_alarm"
